chore(e2e): consolidate redundant specs + split two mega-specs (audit A+C)
Acts on the daily-refactoring-style audit summarised in the
project_e2e_audit_request memory. The full review covered all 85
e2e specs (16 live + 69 mock); this PR lands the two lowest-risk
bundles, A (consolidations) and C (mega-spec splits). B (selective
trims of files-path-url / notifications history / router-nav table-
driven / chatinput-attach counter) and D (collection state-store
consolidation) are deferred to follow-up PRs so this one stays
reviewable.

## A — Consolidations (4 files dropped, behaviour preserved)

- **e2e/tests/history-filter.spec.ts** → folded into
  e2e/tests/history-panel.spec.ts as a second "session-history
  filter pills" describe block. The agent audit's "delete entirely"
  recommendation was wrong: history-filter had 4 unique tests
  (Unread / Human / close-reopen reset / 3-pill switching) the
  panel didn't cover. The panel's narrower single-Bridge test is
  replaced by history-filter's bridge → scheduler → all sequence.
- **e2e/tests/plugin-launcher.spec.ts** → folded into
  e2e/tests/router-navigation.spec.ts as a new "page routing via
  launcher buttons" describe. The launcher's button-click → URL
  contract sits cleanly alongside the existing direct-goto routing
  matrix.
- **e2e/tests/session-tab-bar.spec.ts** → folded into
  e2e/tests/session-history-side-panel.spec.ts as a "visible
  per-tab info" describe. SessionTabBar and SessionHistoryPanel
  share chat-only chrome and a router-driven mount path; the
  four per-tab tests (label / unread dot / origin glyph / Chat
  button badge persistence) sit next to the panel toggle / chat-only
  unmount tests.
- **e2e/tests/feed-custom-view.spec.ts** → folded into
  e2e/tests/collection-view-config.spec.ts. The single feed
  parity assertion (gear opens the same config modal a
  skill-backed collection gets) is one describe block alongside
  the existing project-collection coverage.
- **e2e/tests/accounting-settings-{advanced-toggle,rename}.spec.ts**
  → consolidated into e2e/tests/accounting-settings.spec.ts.
  Both originals tested the same Settings tab against the same
  fixture surface; two describe blocks share the file without
  any redundancy.

## C — Mega-spec splits (organisational; no behaviour change)

- **e2e-live/tests/wiki-nav.spec.ts** (was 694 LOC, 9 tests across
  navigation + piped-links + 5 lint diagnostics) splits into:
  - wiki-nav.spec.ts (303 LOC) — L-14 / L-15 / L-15b nav
    regressions + L-16 in the surviving serial block
  - wiki-piped-links.spec.ts (137 LOC, new) — L-WIKI-PIPE +
    L-WIKI-LINT-PIPE-CLEAN (PR #1312 / issue #1297, kept paired
    so the renderer fix and its lint counterpart stay together)
  - wiki-lint.spec.ts (277 LOC, new) — the 5 lint diagnostics
    (EMPTY-TARGET / BROKEN / ORPHAN parallel + MISSING / TAG-DRIFT
    in a serial sub-describe for the index-mutating ones)
  - .github/workflows/e2e_live_no_llm.yaml matrix updated with
    the two new shards so CI runs each spec in its own job
- **e2e/tests/accounting-detail-and-shortcuts.spec.ts** (was 419
  LOC, 9 tests across 3 unrelated feature areas) splits into:
  - accounting-reports-row-click.spec.ts (Balance Sheet 3 tests +
    P&L 1 test + period shortcuts)
  - accounting-journal-detail-panel.spec.ts (Journal 5 tests
    covering inline detail panel toggle, edit/cancel, close
    behaviour, dedup invariant, #1161 close-from-edit regression)

## Numbers

| Bucket | Before | After | Δ |
|---|---|---|---|
| e2e (mock) specs   | 69 | 65 | −4 |
| e2e-live specs     | 16 | 17 | +1 |
| Total spec files   | 85 | 82 | −3 |
| LOC (touched files)| ~2,371 | ~2,408 | +37 (split overhead from duplicated setup helpers) |

The LOC count goes slightly UP because the two mega-spec splits
each need their own copy of the shared `setupSession` helper +
imports + describe scaffolding. The wins are organisational:
3 fewer files to navigate, each remaining file is single-purpose,
and the wiki-nav split lets CI run navigation / piped-link /
lint diagnostics in parallel shards (each ~90 s instead of one
~7 min job).

Test counts are unchanged: every test() block in the deleted
files lives in a new home, with identical assertions and fixtures.

## Out of scope (deferred to follow-up PRs)

- B (selective trims): files-path-url 30-case → 5-case, notifications
  history-tests collapse, router-nav 18 → 6 happy paths, chatinput-
  attach counter-pattern. ~510 LOC, needs per-case judgement.
- D (collection state-store consolidation): sort-persist + view-mode-
  persist scattered across present-collection.spec.ts and
  collection-{sort-persist,calendar}.spec.ts. ~265 LOC, requires
  agreeing on which file owns the localStorage contract.
