+++ e2e-live/tests/wiki-nav.spec.ts
@@ -1,6 +1,6 @@
 import { randomUUID } from "node:crypto";
 
-import { type Page, expect, test } from "@playwright/test";
+import { expect, test } from "@playwright/test";
 
 import { ONE_MINUTE_MS } from "../../server/utils/time.ts";
 import {
@@ -13,31 +13,22 @@ import {
   restoreWikiIndex,
 } from "../fixtures/live-chat.ts";
 
+// Wiki navigation regressions — the SPA's [[wikilink]] → /wiki/<slug>
+// pipeline. Piped-link bugs (PR #1312) live in wiki-piped-links.spec.ts;
+// lint-report diagnostics live in wiki-lint.spec.ts. This file keeps
+// the core "click a wikilink and land on the right page" coverage,
+// including the non-ASCII slug fuzzy-resolver regression (#1194) and
+// the index-bullet round-trip (B-23).
+
 const L14_TIMEOUT_MS = ONE_MINUTE_MS;
 const L15_TIMEOUT_MS = ONE_MINUTE_MS;
 const L16_TIMEOUT_MS = ONE_MINUTE_MS;
-// L-WIKI-* (the four #1297 regression scenarios below) all use the
-// shared `ONE_MINUTE_MS` budget directly via `test.setTimeout()`; we
-// don't introduce per-test wrappers because `ONE_MINUTE_MS` is
-// already the named constant they would alias to. L14/L15/L16 keep
-// their original per-test consts above for repo-history continuity
-// (intentionally not refactored in this PR).
 
-// Navigate to the wiki lint report and wait for the body-side h1 to
-// render. Body-scoped because the panel chrome also has its own
-// "Wiki Lint Report" h2 — a top-level `getByRole` would otherwise
-// hit strict-mode violation on two matching elements.
-const navigateToWikiLintReport = async (page: Page): Promise<void> => {
-  await page.goto("/wiki/lint-report");
-  await expect(page.getByTestId("wiki-page-body").getByRole("heading", { name: "Wiki Lint Report" })).toBeVisible();
-};
-
-// L-14 / L-15 / the non-mutating L-WIKI-LINT-* tests each seed
-// their own pages and never touch the shared `data/wiki/index.md`,
-// so they parallelise freely. L-16 and L-WIKI-LINT-MISSING /
-// L-WIKI-LINT-TAG-DRIFT all mutate that single shared index file;
-// they share the inner `describe.serial` block below to keep one
-// test's restore from clobbering another's replace.
+// L-14 / L-15 / L-15b each seed their own pages and never touch the
+// shared `data/wiki/index.md`, so they parallelise freely. L-16
+// mutates that single shared index file, so it lives in the inner
+// `describe.serial` block below — keeping one test's restore from
+// clobbering another's replace.
 test.describe.configure({ mode: "parallel" });
 
 test.describe("wiki navigation (real workspace)", () => {
@@ -270,248 +261,17 @@ test.describe("wiki navigation (real workspace)", () => {
     }
   });
 
-  test("L-WIKI-PIPE: [[slug|alias]] 形式のリンクをクリックすると URL に |alias が混入しない", async ({ page }, testInfo) => {
-    test.setTimeout(ONE_MINUTE_MS);
-    // Covers PR #1312 / issue #1297: pre-fix `wikiSlugify` stripped
-    // `|` as a non-ASCII character and concatenated the right-hand-
-    // side alias's ASCII chars into the slug. Three symptoms all
-    // stemmed from the same bug:
-    //   1. lint flagged every `[[slug|alias]]` link as a broken link
-    //      to a slug like `<slug>-<alias-ascii>.md`
-    //   2. `renderWikiLinks` (frontend) emitted
-    //      `<span data-page="<slug>|<alias>">` so clicking produced
-    //      a URL containing `%7C<alias-encoded>`
-    //   3. the visible link text was the raw slug+alias string
-    //      instead of just the display alias
-    //
-    // Post-fix, parser/resolver/renderer all share `parseWikiLink`
-    // (`src/lib/wiki-page/link.ts`) and split on `|`. This spec
-    // exercises 2 and 3 end-to-end against a live mulmoclaude server;
-    // 1 is covered by `findBrokenLinksInPage — [[slug|alias]]
-    // regression` in `test/lib/wiki-page/test_lint.ts`.
-    //
-    // Same nonce strategy as L-14: each project gets unique slugs so
-    // chromium / webkit don't race, and `finally` cleans up its own
-    // pages even if an earlier run died mid-test.
-    const projectSlug = testInfo.project.name;
-    const nonce = `${Date.now()}-${randomUUID().slice(0, 6)}`;
-    const sourceSlug = `e2e-live-wiki-pipe-source-${projectSlug}-${nonce}`;
-    const targetSlug = `e2e-live-wiki-pipe-target-${projectSlug}-${nonce}`;
-    const targetMarker = `wiki-pipe target body marker ${nonce}`;
-    // Display alias deliberately mixes Japanese and a unique ASCII
-    // token so a regression that re-includes the alias in the URL
-    // would be visually obvious (the ASCII suffix would survive
-    // wikiSlugify and end up appended to the path segment).
-    const aliasAsciiToken = `alias-ascii-token-${nonce}`;
-    const displayAlias = `日本語の表示テキスト ${aliasAsciiToken}`;
-    try {
-      await placeWikiPage(sourceSlug, [`# wiki-pipe source`, ``, `[[${targetSlug}|${displayAlias}]]`, ``].join("\n"));
-      await placeWikiPage(targetSlug, [`# wiki-pipe target`, ``, targetMarker, ``].join("\n"));
-      await navigateToWikiPage(page, sourceSlug);
-
-      // Renderer assertions — `parseWikiLink` must put the slug in
-      // data-page and the alias in the visible text. Pre-fix both
-      // were the whole `slug|alias` inner string.
-      const pipeLink = page.locator(`.wiki-link[data-page="${targetSlug}"]`);
-      await expect(pipeLink, "wiki-link's data-page must be the target slug only").toBeVisible();
-      await expect(pipeLink, "visible text must be the display alias, not the raw slug+alias string").toHaveText(displayAlias);
-      // Negative DOM guard — if the renderer regresses and emits
-      // data-page with `|`, this locator would match (it does not on
-      // the post-fix DOM). The selector tolerates the renderer
-      // putting other wiki-links on the page; it just asserts none
-      // of them contain a literal pipe.
-      await expect(page.locator(`.wiki-link[data-page*="|"]`), "no wiki-link's data-page should contain a literal pipe").toHaveCount(0);
-
-      await pipeLink.first().click();
-      // `expectWikiPageBody` covers URL ends with target slug + body
-      // hydrates + B-24 /chat sentinel. The %7C-leak sentinel is
-      // #1297-specific (route owns it, helper doesn't), so it stays
-      // inline.
-      await expectWikiPageBody(page, targetSlug, targetMarker);
-      await expect(page, "URL must not contain a percent-encoded pipe (regression sentinel for %7C alias leak)").not.toHaveURL(/%7C/);
-    } finally {
-      await removeWikiPage(sourceSlug);
-      await removeWikiPage(targetSlug);
-    }
-  });
-
-  test("L-WIKI-LINT-PIPE-CLEAN: lint レポート画面で [[slug|alias]] が broken link 扱いされない", async ({ page }, testInfo) => {
-    test.setTimeout(ONE_MINUTE_MS);
-    // Covers PR #1312 / issue #1297 from the lint-UI side: pre-fix
-    // `findBrokenLinksInPage` slugified the whole `<slug>|<alias>`
-    // string and emitted false-positive broken-link entries on the
-    // /wiki/lint-report page (`<slug>-<alias-ascii>.md not found`).
-    // Post-fix, the lint resolver shares `parseWikiLink` with the
-    // renderer so the alias suffix never reaches the slug.
-    //
-    // Same nonce strategy as L-14: each project gets unique slugs.
-    // The lint endpoint scans the entire workspace, so collision-
-    // safety comes from the per-test nonce — assertions only check
-    // for the seeded slug substring, never a fixed slug.
-    const projectSlug = testInfo.project.name;
-    const nonce = `${Date.now()}-${randomUUID().slice(0, 6)}`;
-    const sourceSlug = `e2e-live-wiki-lint-clean-source-${projectSlug}-${nonce}`;
-    const targetSlug = `e2e-live-wiki-lint-clean-target-${projectSlug}-${nonce}`;
-    const aliasAsciiToken = `lint-clean-alias-ascii-${nonce}`;
-    const displayAlias = `日本語の表示テキスト ${aliasAsciiToken}`;
-    try {
-      await placeWikiPage(sourceSlug, [`# wiki-lint-clean source`, ``, `[[${targetSlug}|${displayAlias}]]`, ``].join("\n"));
-      await placeWikiPage(targetSlug, [`# wiki-lint-clean target`, ``, `body marker ${nonce}`, ``].join("\n"));
-      await navigateToWikiLintReport(page);
-      // Strict negative: no <li> in the lint output mentions our
-      // source page with "not found" — that would be the pre-fix
-      // false-positive shape. The `:has-text` filter scopes by both
-      // tokens so the assertion fails only when our seeded link
-      // actually got flagged as broken (not on unrelated noise).
-      await expect(
-        page.locator(`li:has-text("${sourceSlug}"):has-text("not found")`),
-        "lint must not flag the seeded [[slug|alias]] link as broken (pre-fix false-positive sentinel)",
-      ).toHaveCount(0);
-      // Equally strict: the alias's ASCII token must never end up in
-      // a slug-form "not found" entry. Pre-fix produced
-      // `<slug>-<alias-ascii>.md not found`; post-fix it can't.
-      await expect(
-        page.locator(`li:has-text("${aliasAsciiToken}"):has-text("not found")`),
-        "alias ASCII token must not surface in any 'not found' lint entry",
-      ).toHaveCount(0);
-    } finally {
-      await removeWikiPage(sourceSlug);
-      await removeWikiPage(targetSlug);
-    }
-  });
-
-  test("L-WIKI-LINT-EMPTY-TARGET: lint レポート画面で bare [[Japanese]] が empty target 診断に出る", async ({ page }, testInfo) => {
-    test.setTimeout(ONE_MINUTE_MS);
-    // Covers PR #1312's new "empty target" diagnostic. Pre-fix
-    // bare `[[Japanese title]]` (or `[[#anchor]]`) collapsed via
-    // `wikiSlugify` into an empty string and was reported as a
-    // broken link to `<empty>.md`, indistinguishable from a real
-    // missing-file regression. Post-fix, the resolver detects the
-    // empty-slug case and emits `→ empty target` so operators can
-    // filter the noise apart from genuine `<slug>.md not found`.
-    //
-    // No target page is seeded — by design the link cannot resolve.
-    // Cleanup only touches the source page.
-    const projectSlug = testInfo.project.name;
-    const nonce = `${Date.now()}-${randomUUID().slice(0, 6)}`;
-    const sourceSlug = `e2e-live-wiki-lint-empty-source-${projectSlug}-${nonce}`;
-    // Bare Japanese-only target — wikiSlugify strips every char and
-    // returns "" so the resolver has no slug to look up. Crucially
-    // the target string contains NO ASCII (no nonce, no hyphen, no
-    // digit) — any surviving ASCII would slip through wikiSlugify
-    // and become a non-empty slug, demoting the diagnostic from
-    // "empty target" back to "<slug>.md not found". Per-test
-    // uniqueness comes from `sourceSlug` (which IS nonce-stamped),
-    // so the `<li>:has-text(sourceSlug)` filter still scopes the
-    // assertion to this run only.
-    const bareJapaneseTarget = "日本語のみのターゲット記号終端タイトル";
-    try {
-      await placeWikiPage(sourceSlug, [`# wiki-lint-empty source`, ``, `[[${bareJapaneseTarget}]]`, ``].join("\n"));
-      await navigateToWikiLintReport(page);
-      // Positive: the seeded link must surface as an "empty target"
-      // entry naming our source file and the bare Japanese token.
-      // Pre-fix would have produced "→ <some-ascii-tail>.md not
-      // found" instead of "→ empty target".
-      await expect(
-        page.locator(`li:has-text("${sourceSlug}"):has-text("${bareJapaneseTarget}"):has-text("empty target")`),
-        "lint must report bare Japanese-only [[…]] as empty target diagnostic",
-      ).toHaveCount(1);
-      // Negative: same source file, no "<slug>.md not found" entry
-      // for this link. If the diagnostic regressed back to broken-
-      // link reporting, this would catch it.
-      await expect(
-        page.locator(`li:has-text("${sourceSlug}"):has-text("${bareJapaneseTarget}"):has-text("not found")`),
-        "empty-target case must not also surface as a 'not found' broken link",
-      ).toHaveCount(0);
-    } finally {
-      await removeWikiPage(sourceSlug);
-    }
-  });
-
-  test("L-WIKI-LINT-BROKEN: lint レポート画面で [[bogus-slug]] が broken link 診断に出る", async ({ page }, testInfo) => {
-    test.setTimeout(ONE_MINUTE_MS);
-    // General sanity: the broken-link diagnostic itself still works
-    // post-fix. Distinct from L-WIKI-LINT-EMPTY-TARGET (Japanese
-    // → empty slug) and L-WIKI-LINT-PIPE-CLEAN (alias must NOT
-    // false-positive). Here a plain ASCII slug references a file
-    // that doesn't exist — the canonical broken-link case the user
-    // would encounter when typoing or deleting a target page.
-    //
-    // No target page is seeded — by design the link cannot resolve.
-    const projectSlug = testInfo.project.name;
-    const nonce = `${Date.now()}-${randomUUID().slice(0, 6)}`;
-    const sourceSlug = `e2e-live-wiki-lint-broken-source-${projectSlug}-${nonce}`;
-    const bogusTargetSlug = `e2e-live-wiki-lint-broken-bogus-${projectSlug}-${nonce}`;
-    try {
-      await placeWikiPage(sourceSlug, [`# wiki-lint-broken source`, ``, `[[${bogusTargetSlug}]]`, ``].join("\n"));
-      await navigateToWikiLintReport(page);
-      // Positive: an entry naming both the source file and the bogus
-      // target's `<slug>.md not found` form must appear. The
-      // `bogusTargetSlug` substring also implicitly checks that the
-      // resolver did NOT slugify it down to an empty string — it
-      // is plain ASCII and must survive verbatim.
-      await expect(
-        page.locator(`li:has-text("${sourceSlug}"):has-text("${bogusTargetSlug}.md not found")`),
-        "lint must report [[bogus-slug]] as broken link with '<slug>.md not found' shape",
-      ).toHaveCount(1);
-    } finally {
-      await removeWikiPage(sourceSlug);
-    }
-  });
-
-  test("L-WIKI-LINT-ORPHAN: lint レポート画面で index.md にない page が orphan 診断に出る", async ({ page }, testInfo) => {
-    test.setTimeout(ONE_MINUTE_MS);
-    // Covers `findOrphanPages` end-to-end. Any `<slug>.md` on disk
-    // that the index does not reference must surface as the
-    // dedicated "Orphan page" diagnostic — pre-fix this was merged
-    // with the broken-link line and operators could not filter the
-    // two cases apart.
-    //
-    // No index mutation is needed: the user's existing
-    // `data/wiki/index.md` will not reference our nonce-stamped
-    // slug, so seeding just the page file produces a deterministic
-    // orphan. That keeps this test parallel-safe and lets it run
-    // alongside the other non-mutating L-WIKI-LINT-* tests above.
-    //
-    // The lint report aggregates every orphan on disk, so other
-    // parallel tests' transiently-seeded source pages (L-WIKI-PIPE
-    // et al.) may also surface here while they run. That is fine:
-    // the assertion is scoped to our unique slug via `:has-text`,
-    // not to a total orphan count.
-    const testLabel = testInfo.title.split(":")[0].trim().toLowerCase();
-    const nonce = `${testLabel}-${Date.now()}-${randomUUID().slice(0, 6)}`;
-    const orphanSlug = `e2e-live-wiki-lint-orphan-${testInfo.project.name}-${nonce}`;
-    try {
-      await placeWikiPage(orphanSlug, [`# wiki-lint-orphan ${nonce}`, ``, `orphan body ${nonce}`, ``].join("\n"));
-      await navigateToWikiLintReport(page);
-      // Diagnostic shape: `**Orphan page**: \`<slug>.md\` exists
-      // but is missing from index.md`. Anchor on all three tokens
-      // (slug + "Orphan page" + "missing from index.md") so the
-      // assertion only matches the dedicated diagnostic and not,
-      // e.g., a future "<slug>.md not found" line that happened
-      // to mention the same slug.
-      await expect(
-        page.locator(`li:has-text("${orphanSlug}.md"):has-text("Orphan page"):has-text("missing from index.md")`),
-        "lint must report the seeded page as an orphan diagnostic",
-      ).toHaveCount(1);
-    } finally {
-      await removeWikiPage(orphanSlug);
-    }
-  });
-
-  // The three diagnostics below all replace `data/wiki/index.md`
-  // for the duration of the test. The shared `replaceWikiIndex`
-  // helper has no internal locking, so they MUST run serially with
-  // respect to each other; without this fence one test's
-  // `restoreWikiIndex` would race against a sibling's
-  // `replaceWikiIndex`, leaving the workspace in a hybrid state.
-  // The outer describe stays parallel — the L-14 / L-15 / L-15b /
-  // L-WIKI-PIPE / L-WIKI-LINT-PIPE-CLEAN / L-WIKI-LINT-EMPTY-TARGET
-  // / L-WIKI-LINT-BROKEN / L-WIKI-LINT-ORPHAN tests above do not
-  // touch index.md and can run alongside this block without
-  // contention. Any future test that mutates index.md MUST move
-  // inside this block (or a sibling serial block).
-  test.describe.serial("wiki index-mutating diagnostics", () => {
+  // L-16 replaces `data/wiki/index.md` for the duration of the test.
+  // The shared `replaceWikiIndex` helper has no internal locking, so
+  // any test that mutates index.md MUST run serially with respect to
+  // others that do the same (currently only L-16 in this file; the
+  // L-WIKI-LINT-MISSING / L-WIKI-LINT-TAG-DRIFT pair lives in its own
+  // wiki-lint.spec.ts file, which is a separate Playwright run).
+  // The outer describe stays parallel — the L-14 / L-15 / L-15b
+  // tests above do not touch index.md and can run alongside this
+  // block without contention. Any future test that mutates index.md
+  // MUST move inside this block.
+  test.describe.serial("wiki index-mutating navigation", () => {
     test("L-16: /wiki index に並んだ entry をクリックすると各ページが 404 にならず開ける", async ({ page }, testInfo) => {
       test.setTimeout(L16_TIMEOUT_MS);
       // Covers B-23: the wiki index used to drop or mis-link entries
@@ -576,118 +336,6 @@ test.describe("wiki navigation (real workspace)", () => {
         if (replacedIndex) await restoreWikiIndex(originalIndex);
         await removeWikiPage(slugA);
         await removeWikiPage(slugB);
-      }
-    });
-
-    test("L-WIKI-LINT-MISSING: lint レポート画面で index.md が参照する未存在 file が missing 診断に出る", async ({ page }, testInfo) => {
-      test.setTimeout(ONE_MINUTE_MS);
-      // Covers `findMissingFiles` end-to-end. An index.md row that
-      // references a slug whose `<slug>.md` does not exist on disk
-      // must surface as "Missing file". This is the symmetric
-      // partner to "Orphan page" — together they let operators
-      // reconcile the index vs. the pages directory.
-      //
-      // We replace `data/wiki/index.md` with a synthetic row that
-      // points at a nonce-stamped slug we never seed. Restoration
-      // happens in `finally` via the standard `replaceWikiIndex` →
-      // `restoreWikiIndex` round trip (see L-16's iter-2 fix for
-      // the null-vs-empty restore semantics this inherits).
-      const testLabel = testInfo.title.split(":")[0].trim().toLowerCase();
-      const nonce = `${testLabel}-${Date.now()}-${randomUUID().slice(0, 6)}`;
-      const bogusSlug = `e2e-live-wiki-lint-missing-${testInfo.project.name}-${nonce}`;
-      // Sentinel page so `pages/` is non-empty when the lint route
-      // runs. `collectLintIssues` (server/api/routes/wiki.ts) short-
-      // circuits with a single "Wiki `pages/` directory does not
-      // exist yet" message when `slugs.size === 0` — bypassing
-      // `findMissingFiles` entirely. Locally this never fires
-      // because the developer workspace usually has user-owned wiki
-      // pages, but CI starts from a fresh workspace; by the time
-      // this serial-block test runs, the parallel block's
-      // transient pages have all been cleaned up and L-16's seeded
-      // pair is also already removed, so pages/ is empty. Seeding a
-      // single sentinel here keeps the guard off the critical path.
-      // The sentinel itself shows up as an Orphan-page diagnostic
-      // (it's not in our synthetic index), which does not collide
-      // with the Missing-file assertion below (different slug,
-      // different diagnostic phrase).
-      const sentinelSlug = `e2e-live-wiki-lint-missing-sentinel-${testInfo.project.name}-${nonce}`;
-      // Bullet-link form is what `parseBulletLinkRow` reads to
-      // recover entry.slug from the href — same shape L-16 uses.
-      const newIndex = ["# Wiki Index", "", `- [${bogusSlug} title](pages/${bogusSlug}.md) — missing-file canary`, ""].join("\n");
-      let originalIndex: string | null = null;
-      let replacedIndex = false;
-      try {
-        await placeWikiPage(sentinelSlug, [`# sentinel ${nonce}`, ``, `keeps data/wiki/pages/ non-empty`, ``].join("\n"));
-        originalIndex = await replaceWikiIndex(newIndex);
-        replacedIndex = true;
-        await navigateToWikiLintReport(page);
-        // Diagnostic shape: `**Missing file**: index.md references
-        // \`<slug>\` but the file does not exist`. Anchor on slug
-        // + "Missing file" + "does not exist" so the assertion is
-        // narrowly scoped to our seeded entry, even if the
-        // workspace has unrelated missing-file diagnostics.
-        await expect(
-          page.locator(`li:has-text("${bogusSlug}"):has-text("Missing file"):has-text("does not exist")`),
-          "lint must report the seeded index row as a missing-file diagnostic",
-        ).toHaveCount(1);
-      } finally {
-        if (replacedIndex) await restoreWikiIndex(originalIndex);
-        await removeWikiPage(sentinelSlug);
-      }
-    });
-
-    test("L-WIKI-LINT-TAG-DRIFT: lint レポート画面で frontmatter tag と index tag の drift が診断に出る", async ({ page }, testInfo) => {
-      test.setTimeout(ONE_MINUTE_MS);
-      // Covers `findTagDrift` end-to-end. A page whose YAML
-      // frontmatter `tags:` differs from the matching index row's
-      // hashtag set must surface as "Tag drift" — pre-fix it
-      // surfaced as a generic warning that operators could not
-      // filter. The page exists (no Missing file noise) AND is in
-      // the index (no Orphan page noise) so the assertion can
-      // anchor narrowly on the drift diagnostic itself.
-      //
-      // Both the page body and the index row need to be seeded.
-      // The shared `data/wiki/index.md` is mutated → serial block.
-      const testLabel = testInfo.title.split(":")[0].trim().toLowerCase();
-      const nonce = `${testLabel}-${Date.now()}-${randomUUID().slice(0, 6)}`;
-      const slug = `e2e-live-wiki-lint-drift-${testInfo.project.name}-${nonce}`;
-      // Tag tokens use only lowercase ASCII + hyphens + digits so
-      // both `parseFrontmatterTags` (post-`cleanTagToken`) and
-      // `extractHashTags` (whose regex allows `\p{L}\p{N}_-`) keep
-      // them verbatim. Distinct tags on each side so the drift
-      // condition is unambiguous (set inequality).
-      const pageTag = `pageonly-${nonce}`;
-      const indexTag = `indexonly-${nonce}`;
-      // YAML block-list frontmatter — js-yaml parses both flow
-      // (`tags: [a, b]`) and block forms, but the block form is
-      // unambiguous and survives a future parseFrontmatter rewrite
-      // that drops flow support.
-      const pageBody = ["---", `tags:`, `  - ${pageTag}`, "---", "", `# ${slug}`, "", `drift body ${nonce}`, ""].join("\n");
-      // Bullet-link row with the index-side tag tucked into the
-      // description via `#<tag>` (extractHashTags reads it from
-      // the description line). Keep the slug in the href so
-      // `parseBulletLinkRow` recovers entry.slug = slug.
-      const newIndex = ["# Wiki Index", "", `- [${slug} title](pages/${slug}.md) — drift canary #${indexTag}`, ""].join("\n");
-      let originalIndex: string | null = null;
-      let replacedIndex = false;
-      try {
-        await placeWikiPage(slug, pageBody);
-        originalIndex = await replaceWikiIndex(newIndex);
-        replacedIndex = true;
-        await navigateToWikiLintReport(page);
-        // Diagnostic shape: `**Tag drift**: \`<slug>.md\`
-        // frontmatter has [pageonly...] but index.md has
-        // [indexonly...]`. Both tag tokens are nonce-stamped so
-        // the same <li> can be uniquely identified by joining the
-        // slug + "Tag drift" + both tokens — no risk of a partial
-        // match landing on an unrelated diagnostic.
-        await expect(
-          page.locator(`li:has-text("${slug}.md"):has-text("Tag drift"):has-text("${pageTag}"):has-text("${indexTag}")`),
-          "lint must report the seeded slug as a tag-drift diagnostic",
-        ).toHaveCount(1);
-      } finally {
-        if (replacedIndex) await restoreWikiIndex(originalIndex);
-        await removeWikiPage(slug);
       }
     });
   });
