+++ e2e/tests/history-panel.spec.ts
@@ -2,7 +2,8 @@
 // - toggle opens / closes the panel
 // - opening the panel triggers a fresh /api/sessions fetch
 // - clicking a session row navigates to /chat/:id
-// - filter bar pills hide non-matching sessions without leaving the panel
+// - filter pill bar: active-class for current pill, per-origin filter
+//   behavior, reset on close-then-reopen
 //
 // Scope-matching note: the panel used to live at /history. That route
 // is gone — the filter bar is now panel-local state, and tests that
@@ -91,15 +92,55 @@ test.describe("session-history side panel", () => {
     await expect(page.getByTestId("session-filter-skill")).toBeVisible();
     await expect(page.getByTestId("session-filter-bridge")).toBeVisible();
   });
+});
 
-  test("clicking a filter hides non-matching sessions without leaving the panel", async ({ page }) => {
-    // Override sessions with origin data
+test.describe("session-history filter pills", () => {
+  test.beforeEach(async ({ page }) => {
+    await mockAllApis(page);
+  });
+
+  test("All is active when the panel first opens", async ({ page }) => {
+    await page.goto("/chat");
+    await page.getByTestId("session-history-toggle-off").click();
+
+    await expect(page.getByTestId(`session-item-${SESSION_A.id}`)).toBeVisible();
+    // Class derived in the component: active pill gets bg-blue-600.
+    await expect(page.getByTestId("session-filter-all")).toHaveClass(/bg-blue-600/);
+  });
+
+  test("clicking Unread highlights the pill", async ({ page }) => {
+    await page.goto("/chat");
+    await page.getByTestId("session-history-toggle-off").click();
+    await expect(page.getByTestId(`session-item-${SESSION_A.id}`)).toBeVisible();
+
+    await page.getByTestId("session-filter-unread").click();
+
+    await expect(page.getByTestId("session-filter-unread")).toHaveClass(/bg-blue-600/);
+    await expect(page.getByTestId("session-filter-all")).not.toHaveClass(/bg-blue-600/);
+  });
+
+  test("Human pill keeps default-origin sessions visible", async ({ page }) => {
+    await page.goto("/chat");
+    await page.getByTestId("session-history-toggle-off").click();
+    await expect(page.getByTestId(`session-item-${SESSION_A.id}`)).toBeVisible();
+
+    await page.getByTestId("session-filter-human").click();
+
+    await expect(page.getByTestId("session-filter-human")).toHaveClass(/bg-blue-600/);
+    // Default-origin sessions (no `origin` field) render as `human`
+    // per `originOf`, so they remain visible under the Human filter.
+    await expect(page.getByTestId(`session-item-${SESSION_A.id}`)).toBeVisible();
+  });
+
+  test("switching filters updates the visible session set", async ({ page }) => {
+    // Cover bridge → scheduler → all transitions, not just one pill.
+    // Subsumes the older single-Bridge case from the panel describe.
     await page.route(urlEndsWith("/api/sessions"), (route: Route) =>
       route.fulfill({
         json: {
           sessions: [
             { ...SESSION_A, origin: "bridge" },
-            { ...SESSION_B }, // no origin = human
+            { ...SESSION_B, origin: "scheduler" },
           ],
           cursor: "v1:0",
           deletedIds: [],
@@ -108,21 +149,39 @@ test.describe("session-history side panel", () => {
     );
 
     await page.goto("/chat");
-    await expect(page.getByText("MulmoClaude")).toBeVisible();
-
     await page.getByTestId("session-history-toggle-off").click();
-    // Both visible initially
     await expect(page.getByTestId(`session-item-${SESSION_A.id}`)).toBeVisible();
     await expect(page.getByTestId(`session-item-${SESSION_B.id}`)).toBeVisible();
 
-    // Click Bridge filter — panel stays open, only bridge sessions remain.
     await page.getByTestId("session-filter-bridge").click();
     await expect(page.getByTestId("session-history-side-panel")).toBeVisible();
     await expect(page.getByTestId(`session-item-${SESSION_A.id}`)).toBeVisible();
     await expect(page.getByTestId(`session-item-${SESSION_B.id}`)).toBeHidden();
 
-    // Click All to reset
-    await page.getByTestId("session-filter-all").click();
+    await page.getByTestId("session-filter-scheduler").click();
+    await expect(page.getByTestId(`session-item-${SESSION_A.id}`)).toBeHidden();
     await expect(page.getByTestId(`session-item-${SESSION_B.id}`)).toBeVisible();
+
+    await page.getByTestId("session-filter-all").click();
+    await expect(page.getByTestId(`session-item-${SESSION_A.id}`)).toBeVisible();
+    await expect(page.getByTestId(`session-item-${SESSION_B.id}`)).toBeVisible();
+  });
+
+  test("closing and reopening the panel resets the filter to All", async ({ page }) => {
+    // Panel-local filter state on purpose: reopening should land you
+    // on the familiar "everything" view rather than remembering the
+    // last narrow filter, which can be surprising after a break.
+    await page.goto("/chat");
+    await page.getByTestId("session-history-toggle-off").click();
+    await page.getByTestId("session-filter-unread").click();
+    await expect(page.getByTestId("session-filter-unread")).toHaveClass(/bg-blue-600/);
+
+    // Close (dock toggle in panel header).
+    await page.getByTestId("session-history-toggle-on").click();
+    await expect(page.getByTestId("session-history-side-panel")).toBeHidden();
+
+    // Reopen.
+    await page.getByTestId("session-history-toggle-off").click();
+    await expect(page.getByTestId("session-filter-all")).toHaveClass(/bg-blue-600/);
   });
 });
